Replaced integer millisecond timing with double second. Reduces judder.
git-svn-id: https://s.snth.net/svn/rlk/thumb@3511 dcf3a027-e0e4-0310-89c9-ab9954db6d00

+++ include/app-event.hpp
@@ -145,7 +145,7 @@ namespace app
         };
         struct tick_data_t
         {
-            int dt;
+            double dt;
         };
 
         // Data union
@@ -170,7 +170,7 @@ namespace app
         event *mk_key   (int, int, int, bool);
         event *mk_axis  (int, int, double);
         event *mk_button(int, int, bool);
-        event *mk_tick  (int);
+        event *mk_tick  (double);
         event *mk_draw  ();
         event *mk_swap  ();
         event *mk_user  (long long);
